Better shot state management. Only allows local player to fire up to variable shotSlots, which is now a Byte (since the Byte's range is the max supported). Shots also expire according to variable reloadTime, meaning that just because a shot ends, doesn't necessarily mean you have the available shots to shoot (this is the shot reload timer).

+++ AngryTanks.Common/VariableDatabase.cs
@@ -142,7 +142,7 @@ namespace AngryTanks.Common
             AddVariable("shotRange",
                         "Range of shots", 350f, typeof(Single));
             AddVariable("shotSlots",
-                        "Number of shot slots", 5, typeof(UInt16));
+                        "Number of shot slots", 5, typeof(Byte));
             AddVariable("shotSpeed",
                         "Speed of shots", 50f, typeof(Single));
             AddVariable("tankAngVel",
